+++ lib/std/body.c
@@ -534,7 +534,7 @@ void flee()
 }
 
 // Overriden to add coins weight
-varargs int query_capacity(string relation)
+varargs float query_capacity(string relation)
 {
    float cap = ::query_capacity(relation);
    // int coin_weight_kg = to_int(MONEY_D->coin_weight(query_money()));
